@@ -11,12 +11,12 @@ import {
     ListItemIcon,
     ListItemText,
     Toolbar,
-    Typography
-} from "@material-ui/core";
-import MenuIcon from '@material-ui/icons/Menu';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import DashboardIcon from '@material-ui/icons/Dashboard';
-import {createMuiTheme, makeStyles, ThemeProvider} from "@material-ui/core/styles";
+    Typography,
+} from "@mui/material";
+import MenuIcon from '@mui/icons-material/Menu';
+import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
+import DashboardIcon from '@mui/icons-material/Dashboard';
+import makeStyles from '@mui/styles/makeStyles';
 import clsx from "clsx";
 import LeftContentController from "./components/LeftContent/LeftContentController";
 import qs from 'qs'
@@ -25,98 +25,99 @@ import {connect} from "react-redux";
 import {changeTab, changeUser} from "./redux/actions";
 import {getTabFromID, TABS} from "./constants/model";
 import PlanToolbar from "./components/PlanToolbar";
-import SearchIcon from '@material-ui/icons/Search';
+import SearchIcon from '@mui/icons-material/Search';
 import {logInUser} from "./model/networking";
 
 //https://faizanv.medium.com/authentication-for-your-react-and-express-application-w-json-web-tokens-923515826e0
 
 
-
-
-
 const drawerWidth = 240;
 
-const useStyles = makeStyles((theme) => ({
-    root: {
-        display: 'flex',
-    },
-    toolbar: {
-        minHeight: '80px',
-        paddingRight: 24, // keep right padding when drawer closed
-    },
-    toolbarIcon: {
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'flex-end',
-        padding: '0 8px',
-        ...theme.mixins.toolbar,
-    },
-    appBar: {
-        zIndex: theme.zIndex.drawer + 1,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen,
-        }),
-    },
-    appBarShift: {
-        marginLeft: drawerWidth,
-        width: `calc(100% - ${drawerWidth}px)`,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    },
-    menuButton: {
-        marginRight: 36,
-    },
-    menuButtonHidden: {
-        display: 'none',
-    },
-    title: {
-        flexGrow: 1,
-    },
-    drawerPaper: {
-        position: 'relative',
-        whiteSpace: 'nowrap',
-        width: drawerWidth,
-        transition: theme.transitions.create('width', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    },
-    drawerPaperClose: {
-        overflowX: 'hidden',
-        transition: theme.transitions.create('width', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen,
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up('sm')]: {
-            width: theme.spacing(9),
+
+const useStyles = makeStyles((theme) => {
+    console.log(theme)
+    return {
+        root: {
+            display: 'flex',
         },
-    },
-    appBarSpacer: {
-        marginTop: '80px'
-    },
-    content: {
-        flexGrow: 1,
-        height: '100vh',
-        overflow: 'auto',
-    },
-    container: {
-        paddingTop: theme.spacing(4),
-        paddingBottom: theme.spacing(4),
-    },
-    paper: {
-        padding: theme.spacing(2),
-        display: 'flex',
-        overflow: 'auto',
-        flexDirection: 'column',
-    },
-    fixedHeight: {
-        height: 240,
+        toolbar: {
+            minHeight: '80px',
+            paddingRight: 24, // keep right padding when drawer closed
+        },
+        toolbarIcon: {
+            display: 'flex',
+            alignItems: 'center',
+            justifyContent: 'flex-end',
+            padding: '0 8px',
+            ...theme.mixins.toolbar,
+        },
+        appBar: {
+            zIndex: theme.zIndex.drawer + 1,
+            transition: theme.transitions.create(['width', 'margin'], {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.leavingScreen,
+            }),
+        },
+        appBarShift: {
+            marginLeft: drawerWidth,
+            width: `calc(100% - ${drawerWidth}px)`,
+            transition: theme.transitions.create(['width', 'margin'], {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.enteringScreen,
+            }),
+        },
+        menuButton: {
+            marginRight: 36,
+        },
+        menuButtonHidden: {
+            display: 'none',
+        },
+        title: {
+            flexGrow: 1,
+        },
+        drawerPaper: {
+            position: 'relative',
+            whiteSpace: 'nowrap',
+            width: drawerWidth,
+            transition: theme.transitions.create('width', {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.enteringScreen,
+            }),
+        },
+        drawerPaperClose: {
+            overflowX: 'hidden',
+            transition: theme.transitions.create('width', {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.leavingScreen,
+            }),
+            width: theme.spacing(7),
+            [theme.breakpoints.up('sm')]: {
+                width: theme.spacing(9),
+            },
+        },
+        appBarSpacer: {
+            marginTop: '80px'
+        },
+        content: {
+            flexGrow: 1,
+            height: '100vh',
+            overflow: 'auto',
+        },
+        container: {
+            paddingTop: theme.spacing(4),
+            paddingBottom: theme.spacing(4),
+        },
+        paper: {
+            padding: theme.spacing(2),
+            display: 'flex',
+            overflow: 'auto',
+            flexDirection: 'column',
+        },
+        fixedHeight: {
+            height: 240,
+        }
     }
-}));
+});
 
 const App = (props) => {
     const [open, setOpen] = React.useState(false);
@@ -141,11 +142,6 @@ const App = (props) => {
         }
     })
 
-    const theme = createMuiTheme({
-        palette: {
-            type: "light"
-        }
-    });
 
     const onClickTabItem = (id) => {
         props.changeTab(getTabFromID(id))
@@ -159,60 +155,60 @@ const App = (props) => {
                 </ListItemIcon>
                 <ListItemText primary="Search"/>
             </ListItem>
-            <ListItem button onClick={() => onClickTabItem(TABS.PLAN.id)}>
-                <ListItemIcon>
-                    <DashboardIcon/>
-                </ListItemIcon>
-                <ListItemText primary="Current Plan"/>
-            </ListItem>
+            {props.user &&
+                <ListItem button onClick={() => onClickTabItem(TABS.PLAN.id)}>
+                    <ListItemIcon>
+                        <DashboardIcon/>
+                    </ListItemIcon>
+                    <ListItemText primary="Current Plan"/>
+                </ListItem>
+            }
+
         </div>
     )
 
     return (
-        <ThemeProvider theme={theme}>
-            <div className={classes.root}>
-                <AppBar position="absolute" className={clsx(classes.appBar, open && classes.appBarShift)}>
-                    <Toolbar className={classes.toolbar}>
-                        <IconButton edge='start'
-                                    onClick={handleDrawerOpen}>
-                            <MenuIcon/>
-                        </IconButton>
-                        <Typography variant="h6" noWrap className={classes.title}>
-                            {props.tab.title}
-                        </Typography>
-                        {props.user && <PlanToolbar/>}
-                        <AuthToolbar/>
-                    </Toolbar>
-                </AppBar>
-                <Drawer
-                    variant="permanent"
-                    classes={{
-                        paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
-                    }}
-                    open={open}>
-                    <div style={{marginBottom: '32px'}}>
-                        <IconButton onClick={handleDrawerClose}>
-                            <ChevronLeftIcon/>
-                        </IconButton>
+        <div className={classes.root}>
+            <AppBar position="absolute" className={clsx(classes.appBar, open && classes.appBarShift)}>
+                <Toolbar className={classes.toolbar}>
+                    <IconButton edge='start' onClick={handleDrawerOpen} size="large">
+                        <MenuIcon/>
+                    </IconButton>
+                    <Typography variant="h6" noWrap className={classes.title}>
+                        {props.tab.title}
+                    </Typography>
+                    {props.user && <PlanToolbar/>}
+                    <AuthToolbar/>
+                </Toolbar>
+            </AppBar>
+            <Drawer
+                variant="permanent"
+                classes={{
+                    paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
+                }}
+                open={open}>
+                <div style={{marginBottom: '32px'}}>
+                    <IconButton onClick={handleDrawerClose} size="large">
+                        <ChevronLeftIcon/>
+                    </IconButton>
+                </div>
+                <Divider/>
+                <List>{mainListItems}</List>
+            </Drawer>
+            <main className={classes.content}>
+                <div className={classes.appBarSpacer}/>
+                <div id="body">
+                    <LeftContentController>
+
+                    </LeftContentController>
+                    <div id="scheduleHalf">
+                        <Schedule/>
                     </div>
-                    <Divider/>
-                    <List>{mainListItems}</List>
-                </Drawer>
-                <main className={classes.content}>
-                    <div className={classes.appBarSpacer}/>
-                    <div id="body">
-                        <LeftContentController>
-
-                        </LeftContentController>
-                        <div id="scheduleHalf">
-                            <Schedule/>
-                        </div>
-                    </div>
-                </main>
+                </div>
+            </main>
 
 
-            </div>
-        </ThemeProvider>
+        </div>
     );
 }
 const mapStateToProps = (state) => {
