@@ -1,6 +1,6 @@
 import React from 'react'
-import {List} from "@material-ui/core";
-import {makeStyles} from "@material-ui/core/styles";
+import {List} from "@mui/material";
+import makeStyles from '@mui/styles/makeStyles';
 import CourseListItem from "./CourseListItem";
 import {connect} from "react-redux";
 import {selectedNewPlanWithUpdate} from "../../redux/actions";
